Fix initial timeout error
On VirtualBox 5.0 the initial vagrant up would fail due to a timeout
error. This was because in the VM, the network adapter was not set to
cable, resulting in a exhaustive search for a DHCP server. This search
was longer than the timeout set in the Vagrantfile.

This commit fixes that issue, by setting the connection to cable. This
is also fixed by upgrading to VirtualBox 5.1 or above. For further
information take a look at [this thread on github](https://github.com/chef/bento/issues/682).