Automatically ssh into Lara directory
Usually if you ssh into the vagrant VM, you want to run some artisan or
npm command. This means that you still have to do `cd Code/Lara` each
time you login to the VM. Adding a `cd` command to the `.bashrc` of the
VM enables the developer to go straight to the right directory.